Worked on the search and add forms to include the Music Library number, the composer, and tweaked the year last played.
Starting to rework the way the search controller handles searching for pieces. It is changing from handling all of the form input data to only taking the non-null (ones that have value) and sending them to the model.

From there the model will iterate through the array passed by the controller. Each time it will add an where statement with the key and value.

This way no matter how big the search criteria is, all of the filled out fields will be looked at while discarding all of the empty fields.