Updated the way that searching works. The controller takes the submitted form and filter through to escape and hold onto the values that are not null (ones the user filled out). Next the model takes the filtered array and adds like statements for each key that is in the array (ex. title, year last played, arranger, etc.).
This consolidates the different search functions in earlier revisions. All possible combinations of search criteria should work, I have not stressed tested it yet.

The second major fix would be dealing with the view of a piece's information. The text boxes now auto fill in with the existing information. The rest of the information boxes need to be added (the music library number, the last year played, and the composer). The style sheets are having an issue where it is not being rendered.