Edit Feature Inclusion and Tweaks
The edit controller and view were added, this handles the ability to update piece information for a selected piece. The model method still has to be implemented.

If the type is not changed, then it is filtered out due to the not change default option having an assigned value of null.

An issue dealing with the directory structure and CSS files was also fixed. An absolute path was used, but the base_url method needed to be used so the path did not include the controller and method at the beginning.