Search and Result Updates
A bug effecting the results of searches has been fixed so that the type of the piece being searching for is taken into account.

The result screen now has the table headers and results align properly.

Along with this, the assets folder has now been moved out of the public folder. This keeps the contents inaccessible to visitors of the application.