Action List & Security Update
***Security Update****

GitHub Dependable detected a security vulnerability with CodeIgniter 4.1.7 and below. The project is now running and being developed on 4.1.8 of CodeIgniter. Please upgrade the version of CodeIgniter to match this.
*************************

Worked on the update screen functionality. Now the pieces update when the new information is put in and submitted.

Also worked on the actions of the result page. The delete piece functions properly. The last played year update function does not currently update the year last played in the database, When the link is clicked, the screen is left blank due to no redirection occurring.

temp.html is temporary in testing the year last played update button functionality.