Java Script Alert Added in Add Controller
A test alert was added to the add controller and seems to work. What is left to do is for each action that interfaces with the database. Changing the message and keeping a database operation flag will also be implemented.

Checking for errors when database queries return in an error.